feat(vim): partial refactor, in progress
The packer compiled file is more important than I realised and the
reason things weren't being sourced on second runs is because the
compiled file existed. I thought packer would directly source that but
instead it relies on vim's runtimepath auto-sourcing. Becuase of where I
had put it, it wouldn't auto-source properly.

Now that it is, a good chunk of this refactor (moving all configs under
`lua/config` like they are in `zsh`) is irrelevant so they will need
fixing.

Completion is also wonk :/