feat(nvim): add initial implementation of papercolor in lua
Initial implementation of the papercolor theme in lua to take advantage
of the newer features and to set some TreeSitter highlight variables.

Currently the colours are..... odd. But functionally the whole thing
works which is quite nice. I might end up creating an actual repository
for this and make it more widely available, we will see.