fix(mise): fix `mise` config and add a warning not to use it
The maintainer is dismissive of feedback and the project should not be
supported while they act as such. "This warning message is unclear"
got the response "do this, I think it is clear". Obviously you think
it is clear, you are the one who wrote it. Just as clearly, I thought
it was unclear enough to open a ticket about it.

Obviously their opinion is the one that matters as the maintainer, but
I would stop using this if it wouldn't make sharing configs between my
work and personal overly annoying.

However, I will freeze `mise` and never update it unless absolutely
required and suggest you do the same.